requirements.txt
Django==5.1.1
rq==1.16.2
django-rq==2.10.2
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.15.2
django-redis==5.4.0
gunicorn==23.0.0
drf-nested-routers==0.94.1
django-cors-headers==4.4.0
djangorestframework-datatables==0.7.2
pytz>=2024.2
shapely==2.0.6
numpy==1.26.4